# backend/docs/swagger.yaml
definitions:
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
  handler.getAllBookingsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Booking'
        type: array
    type: object
  handler.responseWithId:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  handler.responseWithToken:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  handler.signInInput:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  model.Booking:
    properties:
      date:
        type: string
      id:
        type: integer
      listing_id:
        type: integer
      user_id:
        type: integer
    required:
    - date
    - listing_id
    - user_id
    type: object
  model.User:
    properties:
      id:
        type: integer
      login:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
    required:
    - login
    - name
    - password
    - role
    type: object
info:
  contact: {}
paths:
  /booking/:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllBookingsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get bookings
      tags:
      - booking
  /signIn/:
    post:
      consumes:
      - application/json
      parameters:
      - description: User's login and password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.responseWithToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Sign-in user
      tags:
      - auth
  /signUp/:
    post:
      consumes:
      - application/json
      parameters:
      - description: User's name, login, password and role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.responseWithId'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Sign-up user
      tags:
      - auth
swagger: "2.0"